[clang-tidy] NO.3 bugprone-narrowing-conversions,cppcoreguidelines-narrowing-conversions PART 1 (#61759)

+++ paddle/fluid/distributed/test/brpc_service_dense_sgd_test.cc
@@ -57,7 +57,7 @@ void InitTensorsOnClient(framework::Scope* scope,
   float* x_ptr =
       x_var->mutable_data<float>(framework::DDim({1, rows_numel}), *place);
   for (int64_t i = 0; i < rows_numel; ++i)
-    x_ptr[i] = 1.0 * static_cast<float>(i);
+    x_ptr[i] = static_cast<float>(1.0) * static_cast<float>(i);
 }
 
 void GetDownpourDenseTableProto(
